clean main code (new file handlers)
